sonar duplicated error
