+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/check-category/components/CheckCategoryColumn.tsx
@@ -1,15 +1,9 @@
-import { IconHash, IconCode, IconArrowsExchange } from '@tabler/icons-react';
+import { IconHash, IconCode } from '@tabler/icons-react';
 import { ColumnDef } from '@tanstack/table-core';
 import { RecordTable, TextOverflowTooltip, RecordTableInlineCell } from 'erxes-ui';
 
-import { CategoryItem, CategoryStatus } from '../types/categoryItem';
-
-const STATUS_STYLES: Record<CategoryStatus, string> = {
-  create: 'bg-green-100 text-green-700',
-  update: 'bg-blue-100 text-blue-700',
-  delete: 'bg-red-100 text-red-700',
-  synced: 'bg-gray-100 text-gray-600',
-};
+import { makeSyncStatusColumn } from '../../shared/syncStatusColumn';
+import { CategoryItem } from '../types/categoryItem';
 
 export const checkCategoryColumns: ColumnDef<CategoryItem>[] = [
   {
@@ -32,22 +26,5 @@ export const checkCategoryColumns: ColumnDef<CategoryItem>[] = [
       </RecordTableInlineCell>
     ),
   },
-  {
-    id: 'status',
-    accessorKey: 'status',
-    size: 90,
-    header: () => <RecordTable.InlineHead label="Status" icon={IconArrowsExchange} />,
-    cell: ({ cell }) => {
-      const status = cell.getValue() as CategoryStatus;
-      return (
-        <RecordTableInlineCell>
-          <span
-            className={`inline-flex items-center rounded px-2 py-0.5 text-xs font-medium capitalize ${STATUS_STYLES[status] ?? ''}`}
-          >
-            {status}
-          </span>
-        </RecordTableInlineCell>
-      );
-    },
-  },
+  makeSyncStatusColumn<CategoryItem>(),
 ];
